Create Navigation and put everything into single component folders, also update Card styling
