+++ public/index.html
@@ -9,6 +9,11 @@
       name="description"
       content="Rick and Morty app created using create-react-app and rickandmortyapi.com"
     />
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+<link href="https://fonts.googleapis.com/css2?family=Audiowide&display=swap" rel="stylesheet">  
+    <style>
+      @import url('https://fonts.googleapis.com/css2?family=DotGothic16&display=swap');
+      </style> 
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
