+++ src/Button/Button.js
@@ -1,8 +1,8 @@
 import './Button.css'
 
-export default function Button({ title, onClick }) {
+export default function Button({ title, onClick, disabled, className }) {
   return (
-    <button className="Button" onClick={onClick}>
+    <button disabled={disabled} className={className} onClick={onClick}>
       {title}
     </button>
   )
